requirements.txt
Django>=4.2,<5
mysqlclient>=2.2
django-environ>=0.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
channels>=4.0
channels-redis>=4.0
scikit-learn>=1.2
gunicorn>=20.1
python-dotenv>=1.0
Pillow>=10.0
django-cors-headers>=4.0
whitenoise>=6.5
psycopg2-binary>=2.9
dj-database-url>=2.1